fix: install button not appearing due to deferredPrompt object not being loaded
This situation happens because deferredPrompt is only registered on page load via the 'beforeinstallprompt' event. Since the initial route of the app isn't at the settings page, this means the old implementation would register the event after it has occured, which means the event listener wont be triggered at all.

The solution solves by storing the deferredPrompt object in a zustand store and watching for the beforeinstallprompt event on app load with a context component above the App component level.